+++ src/main/java/drawers/PointShape.java
@@ -1,6 +1,7 @@
 package drawers;
 
 import java.awt.*;
+import java.util.List;
 
 public class PointShape extends Shape {
     @Override
@@ -8,6 +9,14 @@ public class PointShape extends Shape {
         g.setColor(isHighlight ? Color.green : borderColor);
         new StrokeSetter(g, thickness, isMark, 10);
         g.fillOval(xs1, ys1, 8, 8);
+    }
+
+    private String pointsToString(List<Point> points) {
+        StringBuilder sb = new StringBuilder();
+        for (Point point : points) {
+            sb.append("(").append(point.x).append(",").append(point.y).append(") ");
+        }
+        return sb.toString().trim();
     }
 
     @Override
